chore: reorganizes e2e helper functions
