+++ e2e/infra/deployment.go
@@ -1,4 +1,4 @@
-package e2e
+package infra
 
 import (
 	"io/ioutil"
